mac ui: remove menu bar icon hack, position image at runtime
We were using an source icon that had 16x17 dimensions and then
forcing the 16x17 size property at runtime.
This had too many cross-dependencies
and was assuming undocumented behavior of +[NSImage imageNamed:].

Now we store the source image as 16x16 template (the standard) and
create a new NSImage with 16x17 dimensions at runtime.