mac ui: make dynamically generated menu bar icon image a "template"
A "Template" in cocoa programming is an image containing only pixel data that corresponds to black (or varying alpha levels of black). The default NSStatusItem view will only invert colors of the supplied NSImage if it is set as a template. We weren't setting the template property on our generated menu bar icon NSImage.