mount system: don't communicate with webdav server over HTTP
In the past we communicated with the webdav server over HTTP to tell it to stop or disconnect all clients. This poses a slight security risk because this gives any application on the local machine the capability to do the same.

This change adds a "WebdavServerHandle" class that provides an API for the mount system to internally send stop and disconnect messages to the webdav server. Additionally HTTP code in the mount system was removed since it is no longer necessary.

This change also unifies the webdav server thread routine used by Windows and Mac OS X.