mac/SFXAppDelegate: add_mount_to_favorites() should not fail mount
add_mount_to_favorites() can fail sometimes for unknown reasons
(we depend on LSSharedFileListInsertItemURL() which doesn't
have error codes or much documentation) so wrap it in a try/catch.
This isn't a critical operation so it's okay but we'll need
a way in the future to report non-critical exceptions like this.